Make server infrastructure independent of spring reactive
This commit removes the dependency between this Boot starter and the
server infrastructure in spring-reactive, which primary goal is testing.

Servers are now spawning non-daemon threads in order to keep the Boot
application running after the server has started. This removes the need
for the former workaround using Threads and potentially blocking unit
tests.

Jetty and RxNetty dependencies are upgraded.

The ReactiveHttpServerAutoConfiguration now prevents multiple
`ReactiveHttpServerFactory` to be created in case multiple runtime
libraries are in the classpath; for example, Tomcat for the server and
reactor-netty for the HTTP client runtime.

Fixes #4