Support for some server keys in application.properties
This commit refactors the server infrastructure and introduces the
`ReactiveServerProperties`, which partially duplicates Boot's
`ServerProperties` until this is merged into Boot master.

Also, this commit adds `EmbeddedReactiveHttpServerCustomizer` to mutate
the server instance before it is started by Boot.

For now, only `server.port` and `server.address` are supported. Note
that random and default ports are also supported, for example for
testing an application.