Fix repository declaration
For some reason, Maven is lenient when it comes to resolve SNAPSHOT
dependencies: the `repositories` that should be used could be defined in
a child pom. For milestones (or releases not available on Maven Central),
the `repositories` section must be declared at the same level as the
dependency.

Previously, the `repositories` were defined by
`spring-boot-web-reactive-build` while the actual dependency management
section is defined in the "dependencies" pom. This commit moves the
`repositories` section where it belongs.