Override default idle timeout for servlet containers
Some servlet containers enforce a default idle connection timeout.
Tomcat and Jetty are by default configured with a 30sec timeout.

This commit overrides that default with a 30min timeout.

Not that this is only useful for protocols that are not keeping alive
the connection with heartbeat data, e.g. Server Sent Events.

Fixes: gh-44