Mention stacktrace-mode property in README
This property is for now activated by default in all cases, but should
be only activated by devtools in the future; see temporary support in
`ReactorCoreEnvironmentPostProcessor`.

This options is useful in devmode, since it provides stack information
about error signals. But this option has a big performance penalty and
should be disabled by the developer in non-dev scenarii.